Make sure the assembly resolver is always loaded
Fixes #484
