+++ src/GitHub.VisualStudio/GitHubPackage.cs
@@ -17,6 +17,7 @@ using Microsoft.VisualStudio;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
 using Octokit;
+using GitHub.Helpers;
 
 namespace GitHub.VisualStudio
 {
@@ -32,32 +33,12 @@ namespace GitHub.VisualStudio
     [ProvideOptionPage(typeof(OptionsPage), "GitHub for Visual Studio", "General", 0, 0, supportsAutomation: true)]
     public class GitHubPackage : Package
     {
-        // list of assemblies to be loaded from the extension installation path
-        static readonly string[] ourAssemblies =
-        {
-            "GitHub.Api",
-            "GitHub.App",
-            "GitHub.CredentialManagement",
-            "GitHub.Exports",
-            "GitHub.Exports.Reactive",
-            "GitHub.Extensions",
-            "GitHub.Extensions.Reactive",
-            "GitHub.UI",
-            "GitHub.UI.Reactive",
-            "GitHub.VisualStudio",
-            "GitHub.TeamFoundation",
-            "GitHub.TeamFoundation.14",
-            "GitHub.TeamFoundation.15",
-            "GitHub.VisualStudio.UI",
-            "System.Windows.Interactivity"
-        };
 
         readonly IServiceProvider serviceProvider;
-        static bool resolverInitialized;
 
         static GitHubPackage()
         {
-            InitializeAssemblyResolver();
+            AssemblyResolver.InitializeAssemblyResolver();
         }
 
         public GitHubPackage()
@@ -68,14 +49,6 @@ namespace GitHub.VisualStudio
         public GitHubPackage(IServiceProvider serviceProvider)
         {
             this.serviceProvider = serviceProvider;
-        }
-
-        public static void InitializeAssemblyResolver()
-        {
-            if (resolverInitialized)
-                return;
-            AppDomain.CurrentDomain.AssemblyResolve += LoadAssemblyFromRunDir;
-            resolverInitialized = true;
         }
 
         protected override void Initialize()
@@ -89,34 +62,6 @@ namespace GitHub.VisualStudio
 
             foreach (var menu in menus.DynamicMenus)
                 serviceProvider.AddCommandHandler(menu.Guid, menu.CmdId, menu.CanShow, () => menu.Activate());
-        }
-
-        [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute("Microsoft.Reliability", "CA2001:AvoidCallingProblematicMethods")]
-        static Assembly LoadAssemblyFromRunDir(object sender, ResolveEventArgs e)
-        {
-            try
-            {
-                var name = new AssemblyName(e.Name);
-                if (!ourAssemblies.Contains(name.Name))
-                    return null;
-                var path = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
-                var filename = Path.Combine(path, name.Name + ".dll");
-                if (!File.Exists(filename))
-                    return null;
-                return Assembly.LoadFrom(filename);
-            }
-            catch (Exception ex)
-            {
-                var log = string.Format(CultureInfo.CurrentCulture,
-                    "Error occurred loading {0} from {1}.{2}{3}{4}",
-                    e.Name,
-                    Assembly.GetExecutingAssembly().Location,
-                    Environment.NewLine,
-                    ex,
-                    Environment.NewLine);
-                VsOutputLogger.Write(log);
-            }
-            return null;
         }
 
         void IncrementLaunchCount()
